remove AlterSqlBlockRuleStmt
